Introduce agave-fs crate, move file_io and filesystem ops there
