+++ accounts-db/src/accounts_file.rs
@@ -6,12 +6,12 @@ use {
         account_storage::stored_account_info::{StoredAccountInfo, StoredAccountInfoWithoutData},
         accounts_db::AccountsFileId,
         append_vec::{AppendVec, AppendVecError},
-        buffered_reader::RequiredLenBufFileRead,
         storable_accounts::StorableAccounts,
         tiered_storage::{
             error::TieredStorageError, hot::HOT_FORMAT, index::IndexOffset, TieredStorage,
         },
     },
+    agave_fs::buffered_reader::RequiredLenBufFileRead,
     solana_account::AccountSharedData,
     solana_clock::Slot,
     solana_pubkey::Pubkey,
