+++ accounts-db/src/lib.rs
@@ -32,10 +32,7 @@ mod append_vec;
 pub mod blockhash_queue;
 mod bucket_map_holder;
 mod bucket_map_holder_stats;
-mod buffered_reader;
 pub mod contains;
-mod file_io;
-mod io_uring;
 pub mod is_loadable;
 mod is_zero_lamport;
 mod obsolete_accounts;
@@ -53,11 +50,7 @@ pub mod tiered_storage;
 pub mod utils;
 pub mod waitable_condvar;
 
-pub use {
-    buffered_reader::large_file_buf_reader,
-    file_io::{file_creator, set_path_permissions, FileCreator},
-    obsolete_accounts::{ObsoleteAccountItem, ObsoleteAccounts},
-};
+pub use obsolete_accounts::{ObsoleteAccountItem, ObsoleteAccounts};
 
 #[macro_use]
 extern crate solana_metrics;
