+++ fs/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "agave-fs"
+description = "Agave file system utils"
+documentation = "https://docs.rs/agave-fs"
+version = { workspace = true }
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[lib]
+crate-type = ["lib"]
+name = "agave_fs"
+
+[features]
+agave-unstable-api = []
+
+[dependencies]
+libc = { workspace = true }
+log = { workspace = true }
+slab = { workspace = true }
+smallvec = { workspace = true }
+
+[target.'cfg(target_os = "linux")'.dependencies]
+agave-io-uring = { workspace = true }
+io-uring = { workspace = true }
+
+[dev-dependencies]
+rand = { workspace = true }
+tempfile = { workspace = true }
+test-case = { workspace = true }
