+++ fs/src/io_uring/file_creator.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::arithmetic_side_effects)]
+
 use {
     crate::{
         file_io::FileCreator,
