+++ fs/src/io_uring/sequential_file_reader.rs
@@ -1,3 +1,5 @@
+#![allow(clippy::arithmetic_side_effects)]
+
 use {
     super::{
         memory::{FixedIoBuffer, LargeBuffer},
